requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.2.0
certifi==2019.9.11
chardet==3.0.4
colorzero==1.1
darksky-weather==1.5.0
gpiozero==1.5.1
idna==2.8
multidict==4.5.2
olefile==0.46
Pillow==6.2.0
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.1
requests==2.21.0
RPi.GPIO==0.7.0
six==1.12.0
spidev==3.4
tinydb==3.15.0
urllib3==1.24.3
yarl==1.3.0
